feat!: Add a 'removed' field to macra measures
OCD-3135
