+++ chpl/chpl-resources/src/main/resources/errors.properties
@@ -109,6 +109,7 @@ listing.reasonRequired=A reason is required to complete this listing update when
 listing.firstStatusNotActive=The earliest certification status for any listing on the CHPL must be %s.
 listing.noStatusProvided=A certification status must be provided for every listing on the CHPL.
 listing.duplicateStatus=Certification status must not repeat.
+listing.certStatusChange.notAllowed=User %s does not have permission to change certification status of %s from %s to %s.
 listing.missingTestingLab=Testing Lab is required but not found.
 listing.missingCertificationStatus=Certification Status is required but not found.
 listing.missingPracticeType=Practice type is required but was not found.
@@ -165,6 +166,10 @@ listing.criteria.badAdditionalSoftware=Could not process Additional Software '%s
 listing.criteria.metInvalid=0 or 1 is required to indicate whether %s was met.
 listing.criteria.invalidG1MacraMeasure=Certification %s contains invalid G1 Macra Measure: '%s'.
 listing.criteria.invalidG2MacraMeasure=Certification %s contains invalid G2 Macra Measure: '%s'.
+listing.criteria.removedG1MacraMeasureNoIcs=Certification %s may not reference the G1 Macra Measure: '%s' since this listing does not have ICS. The measure has been removed.
+listing.criteria.removedG2MacraMeasureNoIcs=Certification %s may not reference the G2 Macra Measure: '%s' since this listing does not have ICS. The measure has been removed.
+listing.criteria.removedG1MacraMeasure=Certification %s may not reference the G1 Macra Measure: '%s'. The measure has been removed.
+listing.criteria.removedG2MacraMeasure=Certification %s may not reference the G2 Macra Measure: '%s'. The measure has been removed.
 listing.criteria.badG1MacraMeasure=Could not process G1 Macra Measure '%s'.
 listing.criteria.badG2MacraMeasure=Could not process G2 Macra Measure '%s'.
 listing.criteria.duplicateG1MacraMeasure=Certification %s contains duplicate G1 Macra Measure: '%s'.  The duplicates have been removed.
@@ -309,6 +314,7 @@ listing.upload.emptyRows=The file appears to have a header line with no other in
 listing.upload.couldNotParse=Could not read uploaded file: "%s".
 listing.chplProductNumber.notUnique=The expected CHPL Product Number %s must be unique among all other certified products but one already exists with this value.
 listing.chplProductNumber.changedNotUnique=The CHPL Product Number has changed. The new CHPL Product Number "%s" must be unique among all other certified products but one already exists with this value.
+listing.chplProductNumber.systemChangedNotUnique=The CHPL Product Number will be changed from %s to %s. The new CHPL Product Number must be unique among all other certified products but one already exists with this value.
 
 listing.duplicateTestingLab=Listing contains duplicate Testing Lab: '%s'. The duplicates have been removed.
 listing.criteria.duplicateTestTool=Certification %s contains duplicate Test Tool: Name '%s', Version '%s'. The duplicates have been removed.
