+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/permissions/domains/certifiedproduct/CreateFromPendingActionPermissions.java
@@ -2,6 +2,7 @@ package gov.healthit.chpl.permissions.domains.certifiedproduct;
 
 import org.springframework.stereotype.Component;
 
+import gov.healthit.chpl.dto.listing.pending.PendingCertifiedProductDTO;
 import gov.healthit.chpl.permissions.domains.ActionPermissions;
 
 @Component("certifiedProductCreateFromPendingActionPermissions")
@@ -13,13 +14,14 @@ public class CreateFromPendingActionPermissions extends ActionPermissions {
     }
 
     @Override
-    public boolean hasAccess(Object obj) {
-        if (!(obj instanceof Long)) {
+    public boolean hasAccess(final Object obj) {
+        if (!(obj instanceof PendingCertifiedProductDTO)) {
             return false;
         } else if (getResourcePermissions().isUserRoleAdmin()) {
             return true;
         } else if (getResourcePermissions().isUserRoleAcbAdmin()) {
-            Long acbId = (Long) obj;
+            PendingCertifiedProductDTO pendingListing = (PendingCertifiedProductDTO) obj;
+            Long acbId = pendingListing.getCertificationBodyId();
             return isAcbValidForCurrentUser(acbId);
         } else {
             return false;
