+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/pendingListing/reviewer/edition2015/RequiredData2015Reviewer.java
@@ -842,6 +842,23 @@ public class RequiredData2015Reviewer extends RequiredDataReviewer {
                             } else {
                                 pendingMeasureMap.setMacraMeasure(foundMeasure);
                             }
+                        } else if (pendingMeasureMap.getMacraMeasure() == null) {
+                            MacraMeasureDTO foundMeasure = macraDao.getById(pendingMeasureMap.getMacraMeasureId());
+                            if (foundMeasure == null || foundMeasure.getId() == null) {
+                                listing.getErrorMessages().add(
+                                        msgUtil.getMessage("listing.criteria.invalidG1MacraMeasure",
+                                                cert.getNumber(), pendingMeasureMap.getEnteredValue()));
+                            } else {
+                                pendingMeasureMap.setMacraMeasure(foundMeasure);
+                            }
+                        }
+
+                        if (pendingMeasureMap.getMacraMeasure() != null
+                                && (listing.getIcs() == null || !listing.getIcs().booleanValue())
+                                && pendingMeasureMap.getMacraMeasure().getRemoved() != null
+                                && pendingMeasureMap.getMacraMeasure().getRemoved().booleanValue()) {
+                            listing.getErrorMessages().add(msgUtil.getMessage("listing.criteria.removedG1MacraMeasureNoIcs",
+                                    cert.getNumber(), pendingMeasureMap.getMacraMeasure().getValue()));
                         }
                     }
                 }
@@ -859,6 +876,23 @@ public class RequiredData2015Reviewer extends RequiredDataReviewer {
                             } else {
                                 pendingMeasureMap.setMacraMeasure(foundMeasure);
                             }
+                        } else if (pendingMeasureMap.getMacraMeasure() == null) {
+                            MacraMeasureDTO foundMeasure = macraDao.getById(pendingMeasureMap.getMacraMeasureId());
+                            if (foundMeasure == null || foundMeasure.getId() == null) {
+                                listing.getErrorMessages().add(
+                                        msgUtil.getMessage("listing.criteria.invalidG2MacraMeasure",
+                                                cert.getNumber(), pendingMeasureMap.getEnteredValue()));
+                            } else {
+                                pendingMeasureMap.setMacraMeasure(foundMeasure);
+                            }
+                        }
+
+                        if (pendingMeasureMap.getMacraMeasure() != null
+                                && (listing.getIcs() == null || !listing.getIcs().booleanValue())
+                                && pendingMeasureMap.getMacraMeasure().getRemoved() != null
+                                && pendingMeasureMap.getMacraMeasure().getRemoved().booleanValue()) {
+                            listing.getErrorMessages().add(msgUtil.getMessage("listing.criteria.removedG2MacraMeasureNoIcs",
+                                    cert.getNumber(), pendingMeasureMap.getMacraMeasure().getValue()));
                         }
                     }
                 }
