+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/validation/cmsid/CertificationIdTest.java
@@ -3,8 +3,12 @@ package gov.healthit.chpl.validation.cmsid;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.ff4j.FF4j;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.Mockito;
+import org.mockito.MockitoAnnotations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestExecutionListeners;
@@ -17,6 +21,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.github.springtestdbunit.DbUnitTestExecutionListener;
 import com.github.springtestdbunit.annotation.DatabaseSetup;
 
+import gov.healthit.chpl.FeatureList;
 import gov.healthit.chpl.certificationId.Validator;
 import gov.healthit.chpl.certificationId.ValidatorFactory;
 import gov.healthit.chpl.dto.CQMMetDTO;
@@ -38,6 +43,15 @@ public class CertificationIdTest extends TestCase {
 
     @Autowired
     private ValidatorFactory validatorFactory;
+
+    @Autowired
+    private FF4j ff4j;
+
+    @Before
+    public void setup() {
+        MockitoAnnotations.initMocks(this);
+        Mockito.doReturn(true).when(ff4j).check(FeatureList.EFFECTIVE_RULE_DATE);
+    }
 
     @Test
     @Transactional
